Moving circular buffer code in seperate file. Adding new code file to build
